# specifications/architecture/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DigitalBurj Logistics OS API",
    "version": "1.0.0",
    "description": "Versioned REST contract. Money = decimal strings. Commands are idempotent via Idempotency-Key; mutable aggregates use If-Match."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "SERVICE_UNAVAILABLE",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_REQUIRED",
              "NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_KEY_REUSED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "INVALID_STATE_TRANSITION",
              "DUPLICATE",
              "ACCOUNTING_PERIOD_CLOSED",
              "PERIOD_NOT_FOUND",
              "INSUFFICIENT_STOCK",
              "STOCK_ON_HOLD",
              "RELEASE_EVIDENCE_REQUIRED",
              "SEPARATION_OF_DUTIES",
              "OVER_ALLOCATION",
              "JOB_NOT_CLOSABLE",
              "CURRENCY_MISMATCH",
              "DOCUMENT_NOT_CLEAN",
              "WEBHOOK_SIGNATURE_INVALID",
              "AI_TOOL_DENIED",
              "AI_BUDGET_EXCEEDED",
              "BOOKING_OUTCOME_UNKNOWN",
              "QUALIFICATION_REQUIRED",
              "RATE_LIMITED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/operations": {
      "get": {
        "operationId": "getOperationsHealth",
        "tags": [
          "Administration"
        ],
        "summary": "Tenant queue and service operations",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}/status": {
      "post": {
        "operationId": "changeMemberStatus",
        "tags": [
          "Administration"
        ],
        "summary": "Suspend, revoke or reactivate a membership",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "suspended",
                      "revoked"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "status",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}/revoke-sessions": {
      "post": {
        "operationId": "revokeMemberSessions",
        "tags": [
          "Administration"
        ],
        "summary": "Invalidate previously issued access tokens",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}/grants": {
      "post": {
        "operationId": "replaceMemberGrants",
        "tags": [
          "Administration"
        ],
        "summary": "Replace role grants with validated entity and branch scopes",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  },
                  "grants": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 80
                        },
                        "legalEntityId": {
                          "type": "string",
                          "nullable": true,
                          "format": "uuid",
                          "default": null
                        },
                        "branchId": {
                          "type": "string",
                          "nullable": true,
                          "format": "uuid",
                          "default": null
                        }
                      },
                      "required": [
                        "role"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 30
                  }
                },
                "required": [
                  "reason",
                  "grants"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/members": {
      "get": {
        "operationId": "listMembers",
        "tags": [
          "Administration"
        ],
        "summary": "Tenant members and roles",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addMember",
        "tags": [
          "Administration"
        ],
        "summary": "Add member with a role template",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "minLength": 3
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string"
                  },
                  "workspace": {
                    "type": "string",
                    "enum": [
                      "staff",
                      "customer",
                      "agent",
                      "transporter",
                      "driver",
                      "warehouse"
                    ],
                    "default": "staff"
                  },
                  "partyId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "subject",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "listRoles",
        "tags": [
          "Administration"
        ],
        "summary": "Roles and their permissions",
        "description": "Permission: admin.tenant",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-events": {
      "get": {
        "operationId": "listAuditEvents",
        "tags": [
          "Administration"
        ],
        "summary": "Append-only audit log",
        "description": "Permission: audit.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "entityType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "entityId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  }
}